+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## 1.0.0
+## 1.1.0
+* An optional argument isAlwaysInitialize has been added to getCtrl.
+* Added toDict and fromDict method.
 
+## 1.0.0
 * Improved readme and examples.
 
 ## 0.0.1
-
 * Initial release.